chore: move getExepectedParamsForUsers from base.toolset.ts to sdk
